Make ingredient service to divide view and logic
