+++ app/src/main/java/ru/vsu/csf/bakebudget/services/ProductsService.kt
@@ -1,0 +1,4 @@
+package ru.vsu.csf.bakebudget.services
+
+class ProductsService {
+}
